Fix url syntax for default scrape_uri
The mod_status docs suggest the correct syntax, a GET param on the
`server-status` path. but there are few places that reference
`server-status/` (as a directory). Fix up the default for scrape_url
and update the README to use the correct path and param.

Note: the README already includes the correctly URI just below, in the
usage docs.

Docs ref: https://httpd.apache.org/docs/2.4/mod/mod_status.html#machinereadable